+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `Moar.Atom.ensure_existing_atoms/2` to check for a list of previously defined atoms.
 - Add `Moar.Atom.existing_atom?/1` to check if a string has previously been defined as an atom.
 - Add `Moar.Enum.into!/2` which is like `Enum.into/2` but accepts `nil` as its first argument.
+- `Moar.Map.merge/2` now accepts `nil` values.
 
 ## 1.27.0